Add rng_integrity_check column and scaling ODDS
The rng_integrity_check periodically records a random number, and checks
that number against what the ghost used originally. If there's a
mismatch, the debug text turns red. This should be helpful in catching
some of the more subtle changes that break replayability.

One breaking change was that my getFrame() wasn't always returning 0 on
the first update. I've changed it so it sets to 0 on the first update
rather than on newGame();

Also, now the game gets harder as it goes along.